requirements.txt
alembic==1.6.5
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.0.4
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
migrate==0.3.8
Pillow==8.3.1
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.21
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
